*** IMPORTANT NOTE ***
This version marks a major change in gtatools. A lot of classes were moved out of the project and into libnxcommon, hosted on http://code.google.com/p/libnxcommon/. In particular, all classes from libgtaformats/util were moved there, and the ResourceCache class from libgta. Along this were many changes to different files in gtatools, to reflect this major change, as well as to the build system.
libnxcommon is now needed to compile and run any of the gtatools projects, including libgtaformats. To ease development, libnxcommon can be built together with gtatools. See the file README_NXCOMMON in the src directory.

For the list of changes, please see the changelog of libnxcommon rev2.

Other changes than those listed above and in the libnxcommon changelog include:
	- New IMGArchiveHandler system for the new ArchiveHandler system in libnxcommon, to keep accessing IMG archives through the File APIs supported
	- Build system changes (SVN revision detection should work better now)
	- Some preprocessor definitions were moved to libnxcommon in gtaformats/config.cmake.h
	- Some small bug fixes and miscellaneous changes
	- The bulk of changes comes from having to update some include directives, and from updating the GPL headers

This version of gtatools compiles against libnxcommon rev3.